feat(models): Add lastName to copyWith in user models
The `lastName` field has been added to the `copyWith` method for the `Coach`, `AppUser`, and `Swimmer` models.

This allows for the `lastName` property to be updated when creating a copy of these user objects. In the `Coach` model, a bug was also fixed where `lastName` was incorrectly being assigned the value of `this.name`.