feat(models): Add mainEventIds to Swimmer model
This commit introduces a new `mainEventIds` field to the `Swimmer` model.

This field, a list of strings, will be used to store the IDs of a swimmer's main events. It has been added to the constructor, `fromJson`, and `copyWith` methods, with a default empty list.

A minor formatting change was also applied to the `Event` class definition.