refactor(models): Add photoUrl and correct copyWith logic
This commit introduces a new `photoUrl` field to the base `AppUser` model and its subclasses, `Coach` and `Swimmer`. This field will store the URL for a user's profile photo.

The `copyWith` method in the `Coach` model has been corrected to use `this.lastName` as the fallback for the `lastName` field instead of `this.name`.

Additionally, several unnecessary comments have been removed from the `Coach` and `Swimmer` models.